requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
alembic
factory-boy==3.2.0
pytest_asyncio
httpx
pydantic_ssm_settings
boto3
moto
bcrypt